[refactor] repository 네이밍 변경( 공식문서에 따른 data + repository ) 로 변경
